fix: zeroize event and use pair of mpsc channels
Currently the password persists on the heap and can be trivially
obtained by coredumping the soteria process, even long after the
authentication took place.

The conflated event enum is now split into agent/user ones to be used
in a pair of mpsc channels instead of a single broadcast, so it is
removed from the buffer once received. The event enums (including the
password field) are also zeroized when dropped.

Coredump thus no longer contains the password.